split error overlay in 2 files
